transformed y axis to correctly visualize

+++ src/main/kotlin/gcode/GcodeParser.kt
@@ -73,8 +73,8 @@ class GcodeParser(private val gcodeFile: File) : Iterable<GcodeCommand?>, Iterat
     }
 
     companion object {
-        const val maxWidthMM: Double = 40.0
-        const val maxHeightMM: Double = 40.0
+        const val maxWidthMM: Float = 40f
+        const val maxHeightMM: Float = 40f
 
         const val enableMotorsCommand = "E1\n"
         const val enableFansCommand = "F1\n"
